Still WIP
It can't save more than once and it looks really ugly. Should add a GUI to make it customizable without having to tediously code it. 
But this is used with books like Calculus by James Steward where each section has their own own amount of problem sets and it varies for each chapter. so a tool like wheel of names isn't easy to use with books like this.